@@ -1,15 +1,18 @@
 source 'https://rubygems.org'
 
+ruby '3.0.1'
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 5.0.2'
+gem 'rails', '~> 6.1.3'
 # Use Postgresql as the database for Active Record
 gem 'pg'
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0', '>= 5.0.6'
+gem 'sass-rails', '>= 6'
+# Transpile app-like JavaScript. Read more: https://github.com/rails/webpacker
+gem 'webpacker', '~> 5.0'
 # Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
+gem 'uglifier', '>= 4.2.0'
 # Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
+gem 'coffee-rails', '~> 5.0.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 
@@ -20,7 +23,7 @@ gem 'jquery-rails'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 2.0'
 # bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
+gem 'sdoc', '~> 2.2.0', group: :doc
 
 gem 'devise'
 gem 'devise-i18n'
@@ -30,7 +33,7 @@ gem 'activeadmin_addons'
 gem 'active_admin_import', github: "activeadmin-plugins/active_admin_import"
 gem 'inherited_resources', github: 'activeadmin/inherited_resources'
 gem 'aws-sdk-rails'
-gem 'counter_culture', '~> 1.0'
+gem 'counter_culture', '~> 2.9.0'
 gem 'slim'
 gem 'slim-rails'
 gem 'kaminari'
@@ -41,6 +44,7 @@ gem 'peek'
 gem 'peek-rblineprof'
 gem 'dotenv-rails'
 gem 'whenever'
+gem 'bootsnap', '>= 1.1.0', require: false
 
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
@@ -72,7 +76,7 @@ group :development, :test do
 
   gem 'rspec-rails'
   gem 'rspec_rake'
-  gem 'factory_girl_rails'
+  gem 'factory_bot_rails'
   gem 'faker'
   gem 'rails-controller-testing'
 end
