@@ -1,4 +1,4 @@
-class CreateActiveAdminComments < ActiveRecord::Migration
+class CreateActiveAdminComments < ActiveRecord::Migration[6.0]
   def self.up
     create_table :active_admin_comments do |t|
       t.string :namespace
@@ -9,7 +9,7 @@ class CreateActiveAdminComments < ActiveRecord::Migration
       t.timestamps
     end
     add_index :active_admin_comments, [:namespace]
-    add_index :active_admin_comments, [:author_type, :author_id]
+    # add_index :active_admin_comments, [:author_type, :author_id]
     add_index :active_admin_comments, [:resource_type, :resource_id]
   end
 
