@@ -1,4 +1,4 @@
-class CreateUserRoles < ActiveRecord::Migration
+class CreateUserRoles < ActiveRecord::Migration[6.0]
   def change
     create_table :user_roles do |t|
       t.references :user, index: true
