@@ -1,4 +1,4 @@
-class CreateGroups < ActiveRecord::Migration
+class CreateGroups < ActiveRecord::Migration[6.0]
   def change
     create_table :groups do |t|
       t.string :name, null: false
