@@ -1,4 +1,4 @@
-class CreateMonthlyReportTags < ActiveRecord::Migration
+class CreateMonthlyReportTags < ActiveRecord::Migration[6.0]
   def change
     create_table :monthly_report_tags do |t|
       t.integer :monthly_report_id, null: false
