@@ -1,4 +1,4 @@
-class ChangeUserIdOnUserProfiles < ActiveRecord::Migration
+class ChangeUserIdOnUserProfiles < ActiveRecord::Migration[6.0]
   def change
     remove_index :user_profiles, :user_id
     add_index :user_profiles, :user_id, unique: true 
