@@ -1,4 +1,4 @@
-class CreateInquiries < ActiveRecord::Migration
+class CreateInquiries < ActiveRecord::Migration[6.0]
   def change
     create_table :inquiries do |t|
       t.integer :user_id,       null: false, index: true
