@@ -1,4 +1,4 @@
-class CreateAnnouncements < ActiveRecord::Migration
+class CreateAnnouncements < ActiveRecord::Migration[6.0]
   def change
     create_table :announcements do |t|
       t.string :title, null: false
