@@ -1,4 +1,4 @@
-class CreateArticles < ActiveRecord::Migration[5.0]
+class CreateArticles < ActiveRecord::Migration[6.0]
   def change
     create_table :articles do |t|
       t.references :user, index: true, foreign_key: true
