@@ -1,4 +1,4 @@
-class CreateArticleTags < ActiveRecord::Migration[5.0]
+class CreateArticleTags < ActiveRecord::Migration[6.0]
   def change
     create_table :article_tags do |t|
       t.references :tag, null: false
